\name{TCGAhnscPackage}
\alias{TCGAhnscPackage}
\docType{package}
\title{TCGAhnscPackage:  TCGA Head and Neck squamous cell carcinoma data.
}
\description{
Acquired 2-10-15
}
\details{
\tabular{ll}{
Package: \tab TCGAhnsc\cr
Type: \tab Package\cr
Version: \tab 0.99.9\cr
Date: \tab 2010-07-15\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Paul Shannon <pshannon@systemsbiology.org>
Maintainer: Paul Shannon <pshannon@systemsbiology.org>
}
\references{
none yet...
}
\keyword{dynamic}
\seealso{
  NA
}
\examples{
  NA
}
